pyproject.toml
[project]
name = "baker"
version = "0.1.0"
description = "Add your description here"
authors = "CaFeBaBe <whois@kedavra.dev>"
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["egg", "bacon", "sausage", "tomatoes", "Lobster Thermidor"]
requires-python = ">=3.9"
dependencies = [
    "langchain-community>=0.3.8",
    "langchain>=0.3.9",
    "langchain-xai>=0.1.0",
    "langsmith>=0.1.147",
    "beautifulsoup4>=4.12.3",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "ruff>=0.8.1",
]

[dependency-groups]
dev = [
    "langchain-community>=0.3.8",
    "langchain>=0.3.9",
    "langchain-xai>=0.1.0",
    "langsmith>=0.1.147",
    "beautifulsoup4>=4.12.3",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "ruff>=0.8.1",
]
